Both files start from and use common functions in game_functions.py
